requirements.txt
black
isort
python-dotenv

fastapi==0.87.0
uvicorn==0.19.0

celery==5.2.7
redis==4.3.4
flower

SQLAlchemy==1.4.44
psycopg2-binary==2.9.5
flower